% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lancet.R
\name{theme_lancet}
\alias{theme_lancet}
\title{Custom ggplot2 theme for The Lancet Style}
\usage{
theme_lancet()
}
\value{
A ggplot2 theme object.
}
\description{
This function returns a ggplot2 theme that adheres to the style guidelines of The Lancet.
It includes horizontal grid lines, no vertical grid lines, lines for both x and y axes, and a clean, professional look.
The theme also uses the Lancet color palettes for points, lines, bars, and fills.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(x = displ, y = hwy)) +
  geom_point() +
  theme_lancet()
}
